perf: Dashboard is organized by components
